+++ public/js/util.js
@@ -61,7 +61,8 @@ function getAreaStateProperty(areaId, stateId, property) {
 }
 
 function isAbsolutePath(url) {
-	return url.startsWith("http://") || url.startsWith("https://") || url.startsWith("data:image/");
+	return typeof url === "undefined" || url.startsWith("http://") ||
+			url.startsWith("https://") || url.startsWith("data:image/");
 }
 
 function unescapeNewLinesHTML(text) {