+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## TODO
+## 2020-09-29:
+- Fix bug when empty descriptions or urls
 - Update Manual
 
 ## 2020-09-22: